Much work, much result, small changes. Here go:
Added to alphabet section:
#

This means we now go for [ #: | .#. ] where we earlier had
.#. (for now only done for the a:o rule, buit should be generalised.)
The #: is for compounds. 
TODO: Add the # mark to the stems of compounds.

t:s rule: fixed lentä>i3 to lens00i by adding line
+  Vowel [n|l|r] _ [Vowel:0  %^WG:0  %>:0 i3: ] ;
to the t:s rule.

a before i complex of rules:

 "a:0 before Pret and Pl i when rounded root vowel"
Now says:
a:0 <=> (\a) [(u) u|([u|o]) o] Cns: (Cns:) (Cns:) _  [(Dummy:* - %^AE:) %>:0  [i2:|i3:] | [a:0]] ;

"a:o before Pl i and Pret i"
now says:
a:o <=>  [ #: | .#. ] Cns:* ([i|e|a]) [i|e|a] Cns:*  :Cns _ (Dummy: - %^AE:0) %>:0 [i2: | i3: ] ;

 Note: new start of word / compound [#:|.#.] in order to avoid conflict with 3syll words.
Also removed ^AE from the set of possible dummy symbols, in order to avoid conflict
with the a:e rule.

"a:e in passives"
now says:
+Vx:e <=> _ %^AE:0 (%^WG:0) %>:0 t (t) ;

New: we allow for ^WG, and we want a suff border before the t.