Now, two of the geminate rules have a new left context:
  "Gemination Va"
0:Cx <=> [ .#. | #: ] Cns:*   Vowel:* Vowel Cx _ a %>: :a ;
       where Cx in (  k p t s ) ;

  "Gemination VVa"
0:Cx <=> [ .#. | #: ] Cns:* Vowel Cx _ a %>: :a ;
       where Cx in ( h j l m n r v ) ;

What this says is that we have gemination only after 
a main stress syllable (.#. is beginning of word and #: is
compound boundary). 

This gives the correct partitive form kajavaa. 
Without this rule we get incorrect *kajavvaa.

Now, if this really is what we want also for other forms, 
we should add the same left context to all gemination rules.


git-svn-id: https://gtsvn.uit.no/langtech/trunk/langs/fkv@91219 c7155fb1-f0a7-4240-a2fc-2600b6f42f90