Large revision of HMETA rules, primary to fix Pl Ill forms (^V had come through) Most of the changes are cosmetic, but there are substantial ones as well.
The ti:si rule was changed in order to fix Pl Ill of tosi.

I separated the stem vowel instances Vx:0, e:0 (in Pl) and a:0 
in three different rules, they behave differently. Only the Vx one
is actually matching a copied vowel after the h, for the e there is
an i and for the a an o to the right of the h.

Some rules got better names.

All ^AO are now placed to the left of ^HMETA.

All a:o in front of Pl i and Pret i are fused (but the ones interacting
with HMETA are kept separate).