"Genimation VXi" changed: ^E2I removed from one of the contexts, since input does not contain it.
"e:0 in consonant stems and illative plural" changed:
Left context used to be t:s, is now :s, because it shall also cater for
vuossiin, where the second s is not t:s but geminated 0:s.

"a:0 before Pret and Pl i when rounded root vowel" changed:
Added (%^AO:0) to right context.
At present it is better then before, but it introduced an unresolved
conflict with a:o.