Corrected rules:
   "Gemination VXi"
Added ref to ^AO to the left of suff border.

 "a:o before Pl i and Pret i"
Left context was Cns:, this is not general enough, since
the left context in gemination cases is 0:k, so now
this possibiliety is included.

This whole thing fixed the aikkoin forms.
TODO: check the essive, it now gets gemination, 
which yaml says it should not, but ling intuition
says it perhaps should.



git-svn-id: https://gtsvn.uit.no/langtech/trunk/langs/fkv@106525 c7155fb1-f0a7-4240-a2fc-2600b6f42f90