Corrected rules:
   "Gemination VXi"
Added ref to ^AO to the left of suff border.

 "a:o before Pl i and Pret i"
Left context was Cns:, this is not general enough, since
the left context in gemination cases is 0:k, so now
this possibiliety is included.

This whole thing fixed the aikkoin forms.
TODO: check the essive, it now gets gemination, 
which yaml says it should not, but ling intuition
says it perhaps should.