Checkin from a train journey.
Thorough revision: twolc tests down to 1 REJECTED + 1 new REJECTED, but
yaml failures up 40%. 

Work mainly on HMETA & illative, and on Gemination VXi, but also on
adjacent rules. Changed %> ^HMETA to ^HMETA %>, in order to have a 
consistent border symbol %> with diacritics in front of it.
For the stem vowel deletion rule: Removed a and e from the common
rule, since they show different behaviour from the rest (e:0, e:i, a:o, a:0).


git-svn-id: https://gtsvn.uit.no/langtech/trunk/langs/fkv@137459 c7155fb1-f0a7-4240-a2fc-2600b6f42f90