This is a major change:
Rolled back the v2_ta_odd system, now all the v2_odd verbs use the same lexicon, 
both -ta and -tä, and of the infinitive, only -At is outside the lexc stem.

a/ä deletion in past tense is then done by deleting a and ä.
The v2_ta_odd are left for now, to be removed if we go for this.


git-svn-id: https://gtsvn.uit.no/langtech/trunk/langs/fkv@172032 c7155fb1-f0a7-4240-a2fc-2600b6f42f90